Fix up alias-16 support.
After getting in contact with serplord, I now know that the sw alias
loading was correct. Turns out the gl loaders was mostly correct, just a
mistaken subtract rather than add. And with that, I can implement alias-16
support in glsl. better yet, since all the work is done in the loader, the
renderer doesn't know anything about it :) However, I need to create some
16-bit models for testing.

+++ include/QF/GLSL/qf_alias.h
@@ -31,10 +31,12 @@
 #ifndef __QF_GLSL_qf_alias_h
 #define __QF_GLSL_qf_alias_h
 
+#include "QF/GLSL/types.h"
+
 typedef struct aliasvrt_s {
-	short       st[2];
-	short       normal[3];
-	byte        vertex[3];
+	GLshort     st[2];
+	GLshort     normal[3];
+	GLushort    vertex[3];
 } aliasvrt_t;
 
 void R_InitAlias (void);
